requirements.txt
numpy==2.3.0
pandas==2.3.0
scikit-learn==1.7.0
torch==2.7.1
matplotlib==3.10.3
tensorboard==2.19.0
pyyaml==6.0.2
